refactor(logger): move KMP adapters to OneSignalOSCore
